% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_monitoring.R
\name{cluster_relabeler_via_CWM_density}
\alias{cluster_relabeler_via_CWM_density}
\title{Title}
\usage{
cluster_relabeler_via_CWM_density(
  X,
  group_representative_fake_obs,
  old_partition,
  new_partition,
  new_model,
  old_model
)
}
\arguments{
\item{old_model}{}
}
\description{
Title
}
