% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{CWRM_control}
\alias{CWRM_control}
\title{Title}
\usage{
CWRM_control(
  itermax = 100,
  K_steps = 50,
  n_init = 10,
  zero_tol = 1e-16,
  zero_tol_1 = 1e-16,
  zero_tol_2 = 1e-90,
  trace = 0
)
}
\arguments{
\item{trace}{}
}
\description{
Title
}
