% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_step_monitoring_CWRM.R
\name{first_step_monitoring_CWRM}
\alias{first_step_monitoring_CWRM}
\title{Title}
\usage{
first_step_monitoring_CWRM(
  best_alpha_wise,
  relabeler_list,
  what = c("everything", "groups_proportion", "line_plots_parameters", "ARI",
    "prop_doubtful_assignment")
)
}
\arguments{
\item{what}{}
}
\description{
Title
}
