% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_step_monitoring_CWRM.R
\name{second_step_monitoring_CWRM}
\alias{second_step_monitoring_CWRM}
\title{Title}
\usage{
second_step_monitoring_CWRM(
  model_list,
  selected_alpha,
  eta_threshold_ARI = 0.7,
  max_number_of_optimal_sol = 4
)
}
\arguments{
\item{number_of_optimal_sol}{}
}
\description{
Title
}
