% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_monitoring.R
\name{solutions_ranking_identifier_LIGHTER}
\alias{solutions_ranking_identifier_LIGHTER}
\title{Title}
\usage{
solutions_ranking_identifier_LIGHTER(
  solutions_df,
  epsilon_ARI,
  number_of_optimal_sol
)
}
\arguments{
\item{number_of_optimal_sol}{}
}
\description{
Title
}
