% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWRM.R
\name{trim.cwm}
\alias{trim.cwm}
\title{Title}
\usage{
trim.cwm(
  X,
  Y,
  K,
  niter = 20,
  Ksteps = 10,
  maxfactx = 5,
  maxfacty = 5,
  zero.tol = 1e-16,
  zero.tol1 = 1e-16,
  zero.tol2 = 1e-90,
  trace = 0,
  alpha = 0.1,
  init = 5
)
}
\arguments{
\item{init}{}
}
\description{
Title
}
